Get lights and rooms by their name (#31)
This PR mainly allows to find lights and rooms by their name. A pretty
big refactoring was necessary in order to ease the implementation of
this feature.

We introduced a "`Home` Context" that is load at command bootstrap. This
context contains all the resources that are necessary to display the
information to the user:
- bridge_home
- rooms
- devices
- lights
- grouped lights

We hope that this refactoring will ease future developments.